wsi_model/requirements.txt
# WSI Model Requirements
# ======================

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
h5py>=3.8.0
tqdm>=4.65.0

# Image processing
opencv-python>=4.7.0
scikit-image>=0.20.0
Pillow>=9.5.0

# WSI handling
openslide-python>=1.3.0

# Pretrained models
timm>=0.9.0

# Optional: For stain normalization reference implementation
# staintools>=2.1.0

# Note: OpenSlide requires system libraries
# Ubuntu/Debian:
#   apt-get install openslide-tools libopenslide0
#
# macOS:
#   brew install openslide
#
# For GPU support, ensure CUDA is installed and PyTorch is built with CUDA support
